requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-multipart
passlib[bcrypt]
python-jose
Pillow
reportlab
email-validator
psycopg2-binary